Manage sound and buttons display
Added another state to watch when the timer is up. Then if the user wants to reset another timer, clicking on reset will show the start button again. When time is up if the user cliks on reset instead of ok, the sound alert will stop as well. Removed set button and added the logic to set the alarm into the handleStartStop function.